add client data
